Improves the sorting performance.
It no more recreate the whole participation (current and previous) from the state, but creates an "overlay-like" to implement cache\mutation.

on mainnet all-subnets, after warmup (already included onchain validators are fully tracked):

new:
```
Sorting initialization took 45 ms.
Sorting took: 0 ms

Sorting initialization took 50 ms.
Sorting took: 0 ms

Sorting initialization took 57 ms.
Sorting took: 0 ms
```

old:

```
Sorting initialization took 216 ms.
Sorting took: 1 ms

Sorting initialization took 221 ms.
Sorting took: 1 ms

Sorting initialization took 210 ms.
Sorting took: 1 ms
```


## Documentation

- [ ] I thought about documentation and added the `doc-change-required` label to this PR if updates are required.

## Changelog

- [ ] I thought about adding a changelog entry, and added one if I deemed necessary.